Address more copilot's comments

+++ onnxruntime/core/providers/webgpu/webgpu_utils.h
@@ -120,9 +120,9 @@ class SplitKConfig {
   uint32_t GetMaxDimInnerWithSplitK() const;
 
   struct ConfigAtRange {
-    ConfigAtRange(uint32_t max_dim_inner, float rate);
+    ConfigAtRange(uint32_t max_dim_inner, double rate);
     uint32_t max_dim_inner_with_rate = 0;
-    float max_dim_a_outer_multiplies_dim_b_outer_divides_dim_inner = 0.0f;
+    double max_dim_a_outer_multiplies_dim_b_outer_divides_dim_inner = 0.0;
   };
   std::vector<ConfigAtRange> configs_per_dim_inner_range_;
 };
